requirements.txt
gunicorn==20.0.4
Flask==2.2.3
openai==0.27.7
python-dotenv==1.0.0
Requests==2.31.0
itsdangerous>=2.0
